fix for norminette

+++ philosophers/philo/eating.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/31 22:52:16 by hkubo             #+#    #+#             */
-/*   Updated: 2021/09/05 12:02:59 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/09/05 12:03:22 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ int	eating_log(t_info *info, t_philo *philo_info)
 		return (-1);
 	}
 	if (g_die_flag == 0)
-		printf("%ld%03d %d has taken a fork\n", tv.tv_sec, tv.tv_usec / 1000,
+		printf("%ld%03d %d has taken a fork\n", tv.tv_sec, (int)(tv.tv_usec / 1000),
 			philo_info->philo_num);
 	if (gettimeofday(&tv, NULL) == -1)
 	{
